+++ R/aaa0_tinycodet_help.R
@@ -3,8 +3,6 @@
 #'
 #' @description
 #'
-#' Welcome to the 'tinycodet' introduction help page! \cr
-#' \cr
 #' 'tinycodet' adds some functions to help in your coding etiquette. \cr
 #' It primarily focuses on 4 aspects: \cr
 #' \cr